Реструктуризировал проект под запуск и отладку в Docker
