+++ .vscode/launch.json
@@ -1,0 +1,29 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Launch Backend",
+            "type": "go",
+            "request": "launch",
+            "mode": "auto",
+            "program": "${workspaceFolder}/backend/app/main.go",
+            "cwd": "${workspaceFolder}/backend"
+        },
+        {
+            "name": "Attach to Backend in Docker",
+            "type": "go",
+            "request": "attach",
+            "mode": "remote",
+            "remotePath": "/app",  // путь к коду внутри контейнера
+            "port": 2345,          // порт, на котором Delve работает в контейнере
+            "host": "localhost",   // IP хоста для подключения (localhost, так как мы перенаправляем порт)
+//            "localRoot": "${workspaceFolder}/backend",  // путь к коду на локальной машине
+            "cwd": "${workspaceFolder}/backend",
+            "trace": "verbose"     // включить вывод для отладки (можно убрать, если не нужно)
+            
+          }
+    ]
+}
