--- backend/compose.yaml
@@ -1,26 +1,0 @@
-services:
-  web:
-    image: otus-hla-homework
-    container_name: hw_api
-    ports:
-      - 9090:9090
-    depends_on:
-      postgres:
-        condition: service_healthy
-    volumes:
-      - ./config.json:/app/config.json
-
-  postgres:
-    image: postgres:16
-    container_name: hw_pg
-    ports:
-      - 5432:5432
-    environment:
-      - POSTGRES_DB=hw
-      - POSTGRES_USER=app_hw
-      - POSTGRES_PASSWORD=Passw0rd
-    healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U app_hw -d hw"]
-      interval: 10s
-      timeout: 5s
-      retries: 5
