--- backend/internal/config/api.go
@@ -1,6 +1,0 @@
-package config
-
-type APIConfig struct {
-	ServerAddress  string `yaml:"server_address"`
-	ContextTimeout int    `yaml:"context_timeout"`
-}
