--- backend/internal/config/jwt.go
@@ -1,9 +1,0 @@
-package config
-
-import "time"
-
-type JWTConfig struct {
-	PrivateKey  string        `yaml:"private_key"`
-	PublicKey   string        `yaml:"public_key"`
-	TokenExpiry time.Duration `yaml:"token_expiry"` // время жизни токена
-}
